# Conflicts:
#	src/main/java/ohne/name/RespawnHandle.java

+++ src/main/java/ohne/name/RespawnHandle.java
@@ -5,8 +5,10 @@ import net.minecraft.stats.Stats;
 import net.minecraft.world.Container;
 import net.minecraft.world.entity.EntityEquipment;
 import net.minecraft.world.entity.player.Inventory;
+import net.minecraft.world.level.gamerules.GameRules;
 import ohne.name.GUI.CustomChest;
 import ohne.name.GUI.CustomChestMenuProvider;
+import ohne.name.GUI.TotemCustomMenuType;
 
 import java.awt.*;
 
@@ -17,20 +19,26 @@ public class RespawnHandle {
     public RespawnHandle(ServerPlayer player, ServerPlayer newPlayer) {
         this.Player = player;
         this.newPlayer = newPlayer;
-
-        if (!TotemPlayerHandle.IsTotemDead(player)) {
+        newPlayer.getInventory().replaceWith(player.getInventory());
+        if (!TotemPlayerHandle.IsTotemDead(player) && !player.level().getGameRules().get(GameRules.KEEP_INVENTORY)) {
             newPlayer.destroyVanishingCursedItems();
             newPlayer.openMenu(new CustomChestMenuProvider(newPlayer.getStats().getValue(Stats.CUSTOM.get(Stats.DEATHS))));
             if (newPlayer.containerMenu instanceof CustomChest menu) {
                 menu.SetRespawnHandle(this);
             }
             newPlayer.getInventory().replaceWith(new Inventory(newPlayer, new EntityEquipment()));
-        } else if(TotemPlayerHandle.IsTotemDead(player)) {
-            TotemPlayerHandle.getPlayerHandle(player).setInventory(player.getInventory());
+
         }
     }
 
     public void dropItems(Container container) {
-        //container.get
+        Inventory newInventory = new Inventory(Player, new EntityEquipment());
+        for (int i = 0; i < newInventory.getContainerSize(); i++) {
+            if(!container.getItem(i).getOrDefault(TotemCustomMenuType.IS_FROM_CUSTOM_CHEST, false)) {
+                newInventory.setItem(i, container.getItem(i));
+            }
+        }
+        newPlayer.getInventory().replaceWith(newInventory);
+        newPlayer.closeContainer();
     }
 }
